list: initial "bare" partial implementation
Adds initial "bare" partial implementation of the `list` command that
was created mostly for performance testing back in January 2022. Example
of results: listing O(100) workflows took 0.2 sec in Go, 1.5 sec in
Python, and 2.1 sec in Python AppImage.

We can use this "bare" version to run more performance tests (and to
compare to the swagger version that is coming soon.)